Fix Task not Serializable bug and create more organized packages
