+++ src/main/java/tasks/Event.java
@@ -1,6 +1,7 @@
 package tasks;
 
 import java.time.LocalDate;
+import date.DukeDate;
 
 public class Event extends Task {
     private static final long serialVersionUID = 7664438936982546960L;
@@ -9,13 +10,14 @@ public class Event extends Task {
 
     public Event(String description, String from, String to) {
         super(description);
-        this.from = LocalDate.parse(from, super.INPUT_DATE_FORMAT);
-        this.to = LocalDate.parse(to, super.INPUT_DATE_FORMAT);
+        this.from = DukeDate.parseDateString(from);
+        this.to = DukeDate.parseDateString(to);
+
     }
 
     @Override
     public String toString() {
         return String.format("[E]%s (from: %s to: %s)", super.toString(),
-                from.format(super.OUTPUT_DATE_FORMAT), to.format(super.OUTPUT_DATE_FORMAT));
+                DukeDate.convertDateToString(from), DukeDate.convertDateToString(to));
     }
 }
